+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## 0.17.0
 * Fix Windows crash when calling pair APIs with an unknown deviceId
+* Fix Windows crash when calling startScan again after some time
 * Fix Windows warning `Unable to establish connection on channel` when hot restart
 
 ## 0.16.0